+++ tests/test_cli.py
@@ -69,6 +69,28 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     return changelog_file_path
 
 
+@pytest.fixture
+def changelog_without_unreleased(tmpdir):
+    changelog_file_path = os.path.join(tmpdir, "CHANGELOG_without_unreleased.md")
+    with open(changelog_file_path, mode="wt", encoding="utf-8") as file:
+        file.write(
+            """# Changelog
+All notable changes to this project will be documented in this file.
+
+The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
+and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+
+## [Unreleased]
+
+## [1.0.0] - 2017-04-10
+### Deprecated
+- Known issue 1 (1.0.0) 漢字
+- Known issue 2 (1.0.0)
+"""
+        )
+    return changelog_file_path
+
+
 def test_print_help(changelog: str, capsys: pytest.CaptureFixture):
     with pytest.raises(SystemExit) as exc:
         cli(["--help"])
@@ -93,20 +115,8 @@ def test_print_version(changelog: str, capsys: pytest.CaptureFixture):
     assert captured.out.strip() == f"keepachangelog {__version__}"
 
 
-def test_show_release_pretty(changelog: str, capsys: pytest.CaptureFixture):
-    cli(["show", "1.0.0", changelog])
-
-    captured = capsys.readouterr()
-
-    assert captured.err == ""
-    assert (
-        captured.out.strip()
-        == "Deprecated\n  - Known issue 1 (1.0.0) 漢字\r\n  - Known issue 2 (1.0.0)"
-    )
-
-
 def test_show_release_raw(changelog: str, capsys: pytest.CaptureFixture):
-    cli(["show", "1.0.0", changelog, "--raw"])
+    cli(["show", "1.0.0", changelog])
 
     captured = capsys.readouterr()
 
@@ -128,6 +138,19 @@ def test_create_release_automatic_version(
 
     assert captured.err == ""
     assert captured.out.strip() == "2.0.0"
+
+
+def test_create_release_nothing_to_release(
+    changelog_without_unreleased: str, capsys: pytest.CaptureFixture
+):
+    with pytest.raises(SystemExit) as cm:
+        cli(["release", "-f", changelog_without_unreleased])
+    assert cm.value.code == 2
+
+    captured = capsys.readouterr()
+
+    assert captured.err == f"{changelog_without_unreleased} must contains a description of the release content (within Unreleased section)."
+    assert captured.out == ""
 
 
 def test_create_release_specific_version(changelog: str, capsys: pytest.CaptureFixture):